wrap API into web service WIP
