+++ admin/tool/urlpreview/amd/src/get_url_data.js
@@ -1,9 +1,30 @@
-// this script should eventually generate loading animations on page while the url preview is being fetched
+// This file is part of Moodle - http://moodle.org/
+//
+// Moodle is free software: you can redistribute it and/or modify
+// it under the terms of the GNU General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version.
+//
+// Moodle is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+// GNU General Public License for more details.
+//
+// You should have received a copy of the GNU General Public License
+// along with Moodle.  If not, see <http://www.gnu.org/licenses/>.
 
-import {getUrl} from './repository';
+/**
+ * This script fetches the data required from the external service.
+ * It should eventually generate loading animations while it waits for the data.
+ *
+ * @module     tool_urlpreview/get_url_data
+ * @copyright  2024 Team the Z <your@email.com>
+ * @license    http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
+ */
 
-export const getPreviewTemplate = async() => {
-    // insert logic to retrieve url from html page?
-    const response = await getUrl(url);
-    window.console.log(response);
-}
+import {getPreview} from './repository';
+
+export const getPreviewTemplate = async(url) => {
+    const response = await getPreview(url);
+    document.write(response);
+};
