+++ admin/tool/urlpreview/lang/en/tool_urlpreview.php
@@ -36,6 +36,8 @@ $string['urlpreview:usetool'] = 'Use the urlpreview tool';
 $string['privacy:metadata'] = 'The urlpreview tool only displays previews for input urls and does not store user data.';
 $string['deleteunusedpreviews'] = 'Delete unused previews';
 $string['refresholdpreviews'] = 'Refresh old, frequently used previews';
+$string['privacy:metadata'] = 'The urlpreview tool only displays previews for input urls and does not store user data.';
+$string['loading'] = 'url preview generating \'{$a}\'';
